% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_expression}
\alias{get_expression}
\title{Visualizes spatial expression of a gene}
\usage{
get_expression(seuratObj, proj, feature, tissue_csv)
}
\arguments{
\item{seuratObj}{Seurat object.}

\item{proj}{Project name}

\item{feature}{Gene or feature you want to visualize}

\item{tissue_csv}{"tissue_positions.csv" output from space ranger pipeline}
}
\value{
Visualization of gene expression on tissue sample
}
\description{
Visualizes spatial expression of a gene
}
\examples{
get_expression(Obj,"my_project","Ripk1","tissue_positions.csv")
}
