% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{transfer_clusters}
\alias{transfer_clusters}
\title{Takes Seurat object and does dimensional reduction using UMAP and visualizes the cluster assignments on the Visium tissue sample.}
\usage{
transfer_clusters(seuratObj, proj, dimensions, res, tissue_csv, tissue)
}
\arguments{
\item{seuratObj}{Seurat object.}

\item{proj}{Project name}

\item{dimensions}{Dimensionality of the data set. Use Seurat's ElbowPlot or JackStraw function for determining this.}

\item{res}{Resolution}

\item{tissue_csv}{"tissue_positions.csv" output from space ranger pipeline}

\item(tissue){OPTIONAL: If Seurat object contains cells from multiple samples, specify the name of the tissue sample you want to visualize.}
}
\value{
Visualization of cluster labels on tissue sample
}
\description{
Takes Seurat object and does dimensional reduction using UMAP and visualizes the cluster assignments on the Visium tissue sample.
}
\examples{
transfer_clusters(Obj,"my_project",13,0.5,"tissue_positions.csv","tissue_label")
}
